DSIS-2036 feat: Add complete form for adding new Promotorias
Description:
Created a complete form for adding new Promotorias, allowing users to input all necessary information for the creation of a new Promotoria. The form includes fields for all relevant data, such as name, comarca, municipality, and any other required information.

This addition streamlines the process of adding new Promotorias, providing users with a comprehensive and user-friendly interface for inputting data and creating new entries in the system.